# k_admin/admin-web/apps/web-antd/src/api/kadmin/generated/ebook_favorite.ts
// Code generated by kadmin codegen. DO NOT EDIT.
import { request } from '../client';

export interface EbookFavorite {
  ebookId: number;
  userId: number;
}

export interface EbookFavoritePayload {
  userId: number;
}

export interface EbookFavoriteFilters {
  page?: number;
  pageSize?: number;
}

export interface PageResult<T> {
  items: T[];
  page: number;
  pageSize: number;
  total: number;
}

function queryString<T extends object>(filters: T) {
  const params = new URLSearchParams();
  Object.entries(filters).forEach(([key, value]) => {
    if (value !== undefined && value !== '') {
      params.set(key, String(value));
    }
  });
  const query = params.toString();
  return query ? `?${query}` : '';
}

export function getEbookFavoriteList(filters: EbookFavoriteFilters = {}) {
  return request<PageResult<EbookFavorite>>(`/api/ebook-favorites${queryString(filters)}`);
}

export function getEbookFavorite(id: number) {
  return request<EbookFavorite>(`/api/ebook-favorites/${id}`);
}

export function createEbookFavorite(payload: EbookFavoritePayload) {
  return request<EbookFavorite>(`/api/ebook-favorites`, {
    method: 'POST',
    body: JSON.stringify(payload),
  });
}

export function updateEbookFavorite(id: number, payload: EbookFavoritePayload) {
  return request<EbookFavorite>(`/api/ebook-favorites/${id}`, {
    method: 'PUT',
    body: JSON.stringify(payload),
  });
}

export function deleteEbookFavorite(id: number) {
  return request<boolean>(`/api/ebook-favorites/${id}`, { method: 'DELETE' });
}