# k_admin/internal/kadmin/docs/swagger.yaml
basePath: /api
consumes:
- application/json
definitions:
...
  ebook.EbookPayload:
    properties:
      author:
        type: string
      coverUrl:
        type: string
      createBy:
        type: string
      createTime:
        type: string
      delFlag:
        type: integer
      ebookFormat:
        type: string
      ebookName:
        type: string
      ebookSize:
        type: integer
      ebookStatus:
        type: integer
      ebookSubject:
        type: string
      ebookType:
        type: integer
      ebookUrl:
        type: string
      publisher:
        type: string
      remark:
        type: string
      reviewer:
        type: string
      updateBy:
        type: string
      updateTime:
        type: string
      userId:
        type: integer
    type: object
...
  ebook_favorite.EbookFavoritePayload:
    properties:
      userId:
        type: integer
    type: object
...
  kadmin.SwaggerErrorResponse:
    properties:
      code:
        example: 400
        type: integer
      data: {}
      message:
        example: invalid request
        type: string
      msg:
        example: invalid request
        type: string
    type: object
...
  kadmin.SwaggerResponse:
    properties:
      code:
        example: 0
        type: integer
      data: {}
      message:
        example: ok
        type: string
      msg:
        example: ok
        type: string
    type: object
...
info:
  contact: {}
  description: KAdmin 前后端分离接口文档。除登录、刷新令牌和公开配置外，请使用 Bearer Token 调用接口。
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  title: KAdmin API
  version: "1.0"
paths:
...
  /datum/bookmark/favorite:
    post:
      parameters:
      - description: 收藏信息
        in: body
        name: payload
        required: true
        schema:
          type: object
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/kadmin.SwaggerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/kadmin.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/kadmin.SwaggerErrorResponse'
      summary: 收藏书签
      tags:
      - 桌面端活动
...
  /datum/desktop/bookmark/favorite/{userId}/{bookmarkId}:
    delete:
      parameters:
      - description: 用户 ID
        in: path
        name: userId
        required: true
        type: integer
      - description: 书签 ID
        in: path
        name: bookmarkId
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/kadmin.SwaggerResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/kadmin.SwaggerErrorResponse'
      summary: 取消书签收藏
      tags:
      - 桌面端活动
...
  /datum/desktop/favorite/{userId}/{fileId}:
    delete:
      parameters:
      - description: 用户 ID
        in: path
        name: userId
        required: true
        type: integer
      - description: 文件 ID
        in: path
        name: fileId
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/kadmin.SwaggerResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/kadmin.SwaggerErrorResponse'
      summary: 取消收藏
      tags:
      - 桌面端活动
...
  /datum/download/ebook:
    get:
      parameters:
      - description: 电子书 ID
        in: query
        name: ebookId
        required: true
        type: integer
      responses:
        "200":
          description: 电子书文件流（attachment）
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/kadmin.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/kadmin.SwaggerErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/kadmin.SwaggerErrorResponse'
      summary: 下载电子书流（属主；落 ptmj_ebook_download 记录）
      tags:
      - 桌面端电子书
...
  /datum/favorite:
    post:
      parameters:
      - description: 收藏信息
        in: body
        name: payload
        required: true
        schema:
          type: object
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/kadmin.SwaggerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/kadmin.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/kadmin.SwaggerErrorResponse'
      summary: 收藏试卷
      tags:
      - 桌面端活动
  /datum/favorite/{fileId}:
    get:
      parameters:
      - description: 文件 ID
        in: path
        name: fileId
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/kadmin.SwaggerResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/kadmin.SwaggerErrorResponse'
      summary: 查询试卷收藏状态
      tags:
      - 桌面端活动
...
  /ebook-favorites:
    get:
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      - description: 关键词
        in: query
        name: keyword
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "403":
          description: Forbidden
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 查询电子书收藏列表
      tags:
      - 生成业务
    post:
      parameters:
      - description: 电子书收藏信息
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/ebook_favorite.EbookFavoritePayload'
      - description: 幂等键（8-128 位）
        in: header
        name: Idempotency-Key
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "403":
          description: Forbidden
          schema:
            type: object
        "409":
          description: Conflict
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 新增电子书收藏
      tags:
      - 生成业务
  /ebook-favorites/{id}:
    delete:
      parameters:
      - description: 记录 ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "403":
          description: Forbidden
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 删除电子书收藏
      tags:
      - 生成业务
    get:
      parameters:
      - description: 记录 ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "403":
          description: Forbidden
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 获取电子书收藏详情
      tags:
      - 生成业务
    put:
      parameters:
      - description: 记录 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 电子书收藏信息
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/ebook_favorite.EbookFavoritePayload'
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "403":
          description: Forbidden
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 修改电子书收藏
      tags:
      - 生成业务
...
  /ebooks:
    get:
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      - description: 关键词
        in: query
        name: keyword
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "403":
          description: Forbidden
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 查询电子书列表
      tags:
      - 生成业务
    post:
      parameters:
      - description: 电子书信息
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/ebook.EbookPayload'
      - description: 幂等键（8-128 位）
        in: header
        name: Idempotency-Key
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "403":
          description: Forbidden
          schema:
            type: object
        "409":
          description: Conflict
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 新增电子书
      tags:
      - 生成业务
  /ebooks/{id}:
    delete:
      parameters:
      - description: 记录 ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "403":
          description: Forbidden
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 删除电子书
      tags:
      - 生成业务
    get:
      parameters:
      - description: 记录 ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "403":
          description: Forbidden
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 获取电子书详情
      tags:
      - 生成业务
    put:
      parameters:
      - description: 记录 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 电子书信息
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/ebook.EbookPayload'
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "403":
          description: Forbidden
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 修改电子书
      tags:
      - 生成业务
...
produces:
- application/json
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"