# k_admin/internal/kadmin/generated/ebook_favorite/types.go
// Code generated by kadmin codegen. DO NOT EDIT.

package ebook_favorite

// Permission codes for table ptmj_ebook_favorite.
const (
	ListPermission   = "system:ebook_favorite:list"
	CreatePermission = "system:ebook_favorite:create"
	UpdatePermission = "system:ebook_favorite:update"
	DeletePermission = "system:ebook_favorite:delete"
)

// EbookFavorite is the model of table ptmj_ebook_favorite.
type EbookFavorite struct {
	EbookId int64 `json:"ebookId"`
	UserId  int64 `json:"userId"`
}

// EbookFavoritePayload carries the writable fields of EbookFavorite.
type EbookFavoritePayload struct {
	UserId int64 `json:"userId"`
}

// EbookFavoriteFilter carries the list query parameters.
type EbookFavoriteFilter struct {
	Page     int `json:"page"`
	PageSize int `json:"pageSize"`
}

// EbookFavoritePage is a paginated list result.
type EbookFavoritePage struct {
	Items    []EbookFavorite `json:"items"`
	Total    int64           `json:"total"`
	Page     int             `json:"page"`
	PageSize int             `json:"pageSize"`
}